requirements.txt
streamlit>=1.35,<2.0
plotly>=5.18,<6.0
numpy>=1.26,<2.0
scipy>=1.12,<2.0
streamlit-pdf-viewer>=0.0.12
python-docx>=1.1,<2.0
